chore: prepare v0.2.0 release
- add honeycrisp meta-crate (re-exports unimem, acpu, aruminium, rane)
- add unimem README
- enrich acpu, aruminium, rane READMEs (experimental status, hardware docs, performance)
- refine root README (architecture diagram, correct API examples, crates.io badges)
- bump workspace version 0.1.0 → 0.2.0
- add crates.io metadata to all crates (homepage, keywords, categories, exclude)
- fix all clippy warnings (28 resolved)
- fix stale API references in docs (Source::from_text, matmul_f16 return type, rmsnorm)
- add publish = false to bench crates

+++ acpu/src/matrix/tile_f16.rs
@@ -17,15 +17,15 @@ use super::regs::{XRow, YRow};
 // bit 27: skip_z (no accumulate when set)
 
 const fn fma16_first(xr: XRow, yr: YRow, tile: u8) -> u64 {
-    let x_off = (xr.byte_offset() as u64) << 10;
-    let y_off = yr.byte_offset() as u64;
+    let x_off = xr.byte_offset() << 10;
+    let y_off = yr.byte_offset();
     let z = (tile as u64 & 1) << 20;
     x_off | y_off | z | (1 << 27) // skip_z = no accumulate (overwrite)
 }
 
 const fn fma16_acc(xr: XRow, yr: YRow, tile: u8) -> u64 {
-    let x_off = (xr.byte_offset() as u64) << 10;
-    let y_off = yr.byte_offset() as u64;
+    let x_off = xr.byte_offset() << 10;
+    let y_off = yr.byte_offset();
     let z = (tile as u64 & 1) << 20;
     x_off | y_off | z // accumulate
 }
